requirements.txt
configparser
ultralytics
requests
patool
gdown
natsort
Cython
pandas
numpy
matplotlib
pycocotools
Pillow
numpy<1.24
torch==2.2.2
plotly
kaleido==0.1.0.post1
pyyaml